Move the react <head/> component to its own file
